# Conflicts:
#	attendance-admin/src/main/resources/application.yml

+++ attendance-dao/src/main/java/com/jj/stu/attendance/dao/response/attendance/PageAttendanceResponse.java
@@ -20,7 +20,7 @@ public class PageAttendanceResponse {
     /**
      * 总大小
      */
-    private Integer totalSize;
+    private Long totalSize;
 
     /**
      * 出席名单
