+++ attendance-dao/src/main/java/com/jj/stu/attendance/dao/response/clazz/PageClazzResponse.java
@@ -13,5 +13,5 @@ public class PageClazzResponse {
 
     List<Clazz> clazzList;
 
-    Integer totalSize;
+    Long totalSize;
 }
